消除警告。取颜色索引，返回的是 8-bit 的字节。

+++ rgb24to8/cubemapper.hpp
@@ -86,7 +86,7 @@ namespace cubemapper
 #if USE_MAPPER
 			return router[B][G][R];
 #else
-			return get_nearest_color_index(palette, R, G, B);
+			return static_cast<uint8_t>(get_nearest_color_index(palette, R, G, B));
 #endif
 		}
 	};
